Cleaned Up/Renamed Hex => HexInternals
Renamed to save confusion
